pneumonia_detection/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
streamlit
pillow